# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Advanced Multi-modal AI",
    "summary": "A multimodal runtime surface for orchestration, retrieval, quality profiling, provenance, temporal alignment, public-web intake, and video evidence planning.",
    "version": "0.5.0"
  },
  "paths": {
...
    "/v1/catalog/register": {
      "post": {
        "summary": "Register Dataset",
        "operationId": "register_dataset_v1_catalog_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/connectors/register": {
      "post": {
        "summary": "Register Connector Dataset",
        "operationId": "register_connector_dataset_v1_connectors_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connectors/pipeline-ingest": {
      "post": {
        "summary": "Connector Pipeline Ingest",
        "operationId": "connector_pipeline_ingest_v1_connectors_pipeline_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorPipelineIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/recipes/compile": {
      "post": {
        "summary": "Compile Recipe",
        "operationId": "compile_recipe_v1_recipes_compile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCompileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/data/profile": {
      "post": {
        "summary": "Data Profile",
        "operationId": "data_profile_v1_data_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/infer": {
      "post": {
        "summary": "Infer",
        "operationId": "infer_v1_infer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/video/clean": {
      "post": {
        "summary": "Video Clean",
        "operationId": "video_clean_v1_video_clean_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoCleaningRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/benchmarks/smoke": {
      "get": {
        "summary": "Smoke Benchmark",
        "operationId": "smoke_benchmark_v1_benchmarks_smoke_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "adaptive_transformer",
              "title": "Model Id"
            }
          },
          {
            "name": "iterations",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Iterations"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectorPipelineIngestRequest": {
        "properties": {
          "connector": {
            "$ref": "#/components/schemas/ConnectorConfig"
          },
          "dataset_name": {
            "type": "string",
            "minLength": 1,
            "title": "Dataset Name"
          },
          "owner": {
            "type": "string",
            "minLength": 1,
            "title": "Owner"
          },
          "version": {
            "type": "string",
            "minLength": 1,
            "title": "Version"
          },
          "stream_id": {
            "type": "string",
            "minLength": 1,
            "title": "Stream Id"
          },
          "batch_label": {
            "type": "string",
            "title": "Batch Label",
            "default": "default"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id",
            "default": "adaptive_transformer"
          },
          "runtime_mode": {
            "type": "string",
            "enum": [
              "contract",
              "research"
            ],
            "title": "Runtime Mode",
            "default": "contract"
          },
          "target": {
            "type": "string",
            "enum": [
              "embedding",
              "classification",
              "retrieval",
              "translation"
            ],
            "title": "Target",
            "default": "embedding"
          },
          "num_classes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Classes"
          },
          "modality_mappings": {
            "items": {
              "$ref": "#/components/schemas/ConnectorModalityMapping"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 1,
            "title": "Modality Mappings"
          },
          "partition_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Partition Keys"
          },
          "primary_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Primary Keys"
          },
          "baseline_label": {
            "type": "string",
            "title": "Baseline Label",
            "default": ""
          },
          "block_population_drift": {
            "type": "boolean",
            "title": "Block Population Drift",
            "default": true
          },
          "limit": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 256
          },
          "sample_size": {
            "type": "integer",
            "maximum": 25.0,
            "minimum": 1.0,
            "title": "Sample Size",
            "default": 5
          },
          "partition_key_field": {
            "type": "string",
            "title": "Partition Key Field",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "drop_invalid_rows": {
            "type": "boolean",
            "title": "Drop Invalid Rows",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "connector",
          "dataset_name",
          "owner",
          "version",
          "stream_id",
          "modality_mappings"
        ],
        "title": "ConnectorPipelineIngestRequest"
      },
      "ConnectorRegistrationRequest": {
        "properties": {
          "connector": {
            "$ref": "#/components/schemas/ConnectorConfig"
          },
          "dataset_name": {
            "type": "string",
            "minLength": 1,
            "title": "Dataset Name"
          },
          "owner": {
            "type": "string",
            "minLength": 1,
            "title": "Owner"
          },
          "version": {
            "type": "string",
            "minLength": 1,
            "title": "Version"
          },
          "modality": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "video",
              "sensor",
              "tabular"
            ],
            "title": "Modality",
            "default": "tabular"
          },
          "partition_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Partition Keys"
          },
          "primary_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Primary Keys"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "limit": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 256
          },
          "sample_size": {
            "type": "integer",
            "maximum": 25.0,
            "minimum": 1.0,
            "title": "Sample Size",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "connector",
          "dataset_name",
          "owner",
          "version"
        ],
        "title": "ConnectorRegistrationRequest"
      },
...
      "DatasetField": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "dtype": {
            "type": "string",
            "minLength": 1,
            "title": "Dtype"
          },
          "nullable": {
            "type": "boolean",
            "title": "Nullable",
            "default": true
          },
          "semantic_role": {
            "type": "string",
            "title": "Semantic Role",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "dtype"
        ],
        "title": "DatasetField"
      },
      "DatasetRegistrationRequest": {
        "properties": {
          "dataset_name": {
            "type": "string",
            "minLength": 1,
            "title": "Dataset Name"
          },
          "owner": {
            "type": "string",
            "minLength": 1,
            "title": "Owner"
          },
          "version": {
            "type": "string",
            "minLength": 1,
            "title": "Version"
          },
          "modality": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "video",
              "sensor",
              "tabular"
            ],
            "title": "Modality"
          },
          "partition_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Partition Keys"
          },
          "primary_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Primary Keys"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/DatasetField"
            },
            "type": "array",
            "maxItems": 512,
            "minItems": 1,
            "title": "Fields"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "dataset_name",
          "owner",
          "version",
          "modality",
          "fields"
        ],
        "title": "DatasetRegistrationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}